Use Leading-Caps for Datatypes
non-permissive compilation (default) in C++ is not happy when you call
a datatype and a variable the same thing.  This was creating problems
for `status status` and `info *info`.

It would be possible to decorate the member variables with underscores
to distinguish them from the types.  However, that distinction was
already being used for constructor parameters vs. same-named member
variables.

This just makes a change to use leading-caps for datatypes in the
utest namespace, which is a fairly common and readable convention.